Ch: Code refactoring
This commit moves all models into a single file and refactors code
to incoperate the changes
