+++ api/__init__.py
@@ -6,6 +6,7 @@ from flasgger import Swagger
 from flask_script import Manager
 
 from config.config import app_settings
+from models.database_objects import DatabaseObjects
 
 app = Flask(__name__)
 app.config.from_object(app_settings[os.environ.get("APP_ENV")])
@@ -13,6 +14,7 @@ bcrypt = Bcrypt(app)
 jwt = JWTManager(app)
 swagger = Swagger(app)
 manager = Manager(app)
+connection = DatabaseObjects()
 
 #  Register blueprints
 from .product import product as product_bp
